refactor: use new filterEdit and filterList exports for field components
