+++ frontend/components/FieldNumber/FieldNumberFilterEdit.jsx
